re committed Order.java - changed to double

+++ WealthNow/src/com/fdm/wealthnow/common/Order.java
@@ -23,7 +23,7 @@ public class Order {
 	 * Order constructor is to be called to create an OpenOrder
 	 */
 	public Order(Integer user_id, Integer order_id, String currency_code, String order_type, Integer quantity,
-			String stock_symbol, String price_type, Date opening_order_date, Float limit_price, String term,
+			String stock_symbol, String price_type, Date opening_order_date, Double limit_price, String term,
 			String status) {
 		this.user_id = user_id;
 		this.order_id = order_id;
@@ -41,8 +41,8 @@ public class Order {
 	 * Order constructor is to be called to create a ProcessedOrder
 	 */
 	public Order(Integer user_id, Integer order_id, String currency_code, String order_type, Integer quantity,
-			String stock_symbol, String price_type, Date place_order_date, Float limit_price,
-			Date order_completion_date, String status, Float closing_price, Float open_market_price) {
+			String stock_symbol, String price_type, Date place_order_date, Double limit_price,
+			Date order_completion_date, String status, Double closing_price, Double open_market_price) {
 		this.user_id = user_id;
 		this.order_id = order_id;
 		this.currency_code = currency_code;
@@ -129,27 +129,27 @@ public class Order {
 		this.place_order_date = place_order_date;
 	}
 
-	public Float getClosing_price() {
+	public Double getClosing_price() {
 		return closing_price;
 	}
 
-	public void setClosing_price(Float closing_price) {
+	public void setClosing_price(Double closing_price) {
 		this.closing_price = closing_price;
 	}
 
-	public Float getOpen_market_price() {
+	public Double getOpen_market_price() {
 		return open_market_price;
 	}
 
-	public void setOpen_market_price(Float open_market_price) {
+	public void setOpen_market_price(Double open_market_price) {
 		this.open_market_price = open_market_price;
 	}
 
-	public Float getLimit_price() {
+	public Double getLimit_price() {
 		return limit_price;
 	}
 
-	public void setLimit_price(Float limit_price) {
+	public void setLimit_price(Double limit_price) {
 		this.limit_price = limit_price;
 	}
